data-scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
selenium==4.16.0
webdriver-manager==4.0.1
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
validators==0.22.0
colorlog==6.8.0
